UserContext added to ProjectResourceController

+++ src/main/java/faang/school/projectservice/service/resource/ProjectResourceServiceImpl.java
@@ -3,20 +3,24 @@ package faang.school.projectservice.service.resource;
 import faang.school.projectservice.dto.resource.ResourceDto;
 import faang.school.projectservice.exceptions.NotFoundException;
 import faang.school.projectservice.jpa.ResourceRepository;
+import faang.school.projectservice.jpa.TeamMemberJpaRepository;
 import faang.school.projectservice.mapper.ResourceMapper;
 import faang.school.projectservice.model.Project;
 import faang.school.projectservice.model.Resource;
 import faang.school.projectservice.model.ResourceStatus;
 import faang.school.projectservice.model.ResourceType;
+import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.repository.ProjectRepository;
 import faang.school.projectservice.service.s3.AmazonS3Service;
 import faang.school.projectservice.validation.resource.ProjectResourceValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.InputStream;
 import java.math.BigInteger;
+import java.time.LocalDateTime;
 
 @Service
 @RequiredArgsConstructor
@@ -25,13 +29,16 @@ public class ProjectResourceServiceImpl implements ProjectResourceService {
     private final AmazonS3Service amazonS3Service;
     private final ProjectRepository projectRepository;
     private final ResourceRepository resourceRepository;
+    private final TeamMemberJpaRepository teamMemberRepository;
     private final ProjectResourceValidator projectResourceValidator;
     private final ResourceMapper resourceMapper;
 
     @Override
-    public ResourceDto saveFile(long projectId, MultipartFile file) {
+    @Transactional
+    public ResourceDto saveFile(long userId, long projectId, MultipartFile file) {
 
         Project project = projectRepository.getProjectById(projectId);
+        TeamMember teamMember = findTeamMemberByUserIdAndProjectId(userId, projectId);
 
         projectResourceValidator.validateMaxStorageSize(project, file.getSize());
 
@@ -42,10 +49,14 @@ public class ProjectResourceServiceImpl implements ProjectResourceService {
                 .key(key)
                 .name(file.getName())
                 .size(BigInteger.valueOf(file.getSize()))
+                .allowedRoles(teamMember.getRoles())
                 .type(ResourceType.valueOf(file.getContentType()))
-                .project(project)
                 .status(ResourceStatus.ACTIVE)
-                //...
+                .createdBy(teamMember)
+                .updatedBy(teamMember)
+                .createdAt(LocalDateTime.now())
+                .updatedAt(LocalDateTime.now())
+                .project(project)
                 .build();
 
         resource = resourceRepository.save(resource);
@@ -54,14 +65,23 @@ public class ProjectResourceServiceImpl implements ProjectResourceService {
     }
 
     @Override
-    public InputStream getFile(long resourceId) {
-        Resource resource = findById(resourceId);
+    @Transactional(readOnly = true)
+    public InputStream getFile(long userId, long projectId, long resourceId) {
+
+        Resource resource = findResourceById(resourceId);
+        findTeamMemberByUserIdAndProjectId(userId, projectId);
+
         return amazonS3Service.downloadFile(resource.getKey());
     }
 
     @Override
-    public void deleteFile(long resourceId) {
-        Resource resource = findById(resourceId);
+    @Transactional
+    public void deleteFile(long userId, long projectId, long resourceId) {
+
+        Resource resource = findResourceById(resourceId);
+        TeamMember teamMember = findTeamMemberByUserIdAndProjectId(userId, projectId);
+
+        projectResourceValidator.validateDeletePermission(teamMember, resource);
 
         amazonS3Service.deleteFile(resource.getKey());
 
@@ -71,13 +91,19 @@ public class ProjectResourceServiceImpl implements ProjectResourceService {
         resource.setKey(null);
         resource.setSize(null);
         resource.setStatus(ResourceStatus.DELETED);
-        //resource.setUpdatedBy();
+        resource.setUpdatedBy(teamMember);
+        resource.setUpdatedAt(LocalDateTime.now());
 
         resourceRepository.save(resource);
     }
 
-    private Resource findById(Long id) {
+    private Resource findResourceById(long id) {
         return resourceRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Resource with id=" + id + " not found"));
+    }
+
+    private TeamMember findTeamMemberByUserIdAndProjectId(long userId, long projectId) {
+        return teamMemberRepository.findByUserIdAndProjectId(userId, projectId)
+                .orElseThrow(() -> new NotFoundException("TeamMember with userId=" + userId + " and projectId= " + projectId + " not found"));
     }
 }
